requirements.txt
asgiref==3.3.4
Django==3.2.3
django-cors-headers==3.7.0
django-environ==0.4.5
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
greenlet==1.1.0
gunicorn==20.1.0
psycopg2==2.8.6
PyJWT==1.7.1
pytz==2021.1
SQLAlchemy==1.4.17
sqlparse==0.4.1
